WIP: some attempt at abstractions ported from ASP.NET codebase
Not sure this is the right path to go.
